requirements.txt
fake-useragent== 1.5.1
requests==2.32.3
bs4==0.0.2
PyMuPDF==1.24.10
langchain-community==0.3.1
faiss-cpu==1.8.0.post1
sentence_transformers==3.1.1
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.10